requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
environs==11.0.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastapi_cors==0.0.6
Flask==3.0.3
frozenlist==1.4.1
grpclib==0.4.7
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mmh3==4.1.0
modal==0.62.150
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pinecone-text==0.9.0
protobuf==4.25.3
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
sigtools==4.0.1
sniffio==1.3.1
starlette==0.37.2
synchronicity==0.6.7
toml==0.10.2
tqdm==4.66.4
typer==0.12.3
types-certifi==2021.10.8.3
types-requests==2.31.0.20240406
types-toml==0.10.8.20240310
typing_extensions==4.11.0
ujson==5.9.0
unkey.py==0.7.2
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
wget==3.2
yarl==1.9.4
